# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.dealRequest:
    properties:
      count:
        example: 1
        minimum: 1
        type: integer
    required:
    - count
    type: object
  handlers.registerRequest:
    properties:
      username:
        example: alice42
        type: string
    required:
    - username
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'Card-game engine: shoe management, dealing, shuffling, scoring.'
  title: DeckForge API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Returns a fresh JWT for an existing username.
      parameters:
      - description: username
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.registerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: token + user
          schema:
            additionalProperties: true
            type: object
        "400":
          description: bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: username not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log in as an existing user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates an account with a username-only credential and returns
        a signed JWT.
      parameters:
      - description: username (3–15 alphanumeric chars)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: token + user
          schema:
            additionalProperties: true
            type: object
        "400":
          description: validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: username already taken
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
...
  /games/{id}/deal-round:
    post:
      consumes:
      - application/json
      description: 'Server-side atomic deal: deals N cards to every active player
        in seat order. Eliminates N sequential HTTP round-trips and race conditions.
        Auto-ends if shoe is exhausted.'
      parameters:
      - description: game UUID
        in: path
        name: id
        required: true
        type: string
      - description: cards per player
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.dealRequest'
      produces:
      - application/json
      responses:
        "200":
          description: totalDealt + gameEnded
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deal cards to all active players
      tags:
      - dealing
...
  /games/{id}/players:
    get:
      description: Returns all active players sorted by total hand value descending.
        Ties broken by seat order ascending. Scores are intentionally public during
        play — cards remain private.
      parameters:
      - description: game UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: leaderboard array [{playerId, userId, username, seatOrder,
            handValue, cardCount}]
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get the leaderboard
      tags:
      - players
    post:
      description: Adds the authenticated user as a player. If the game is IN_PROGRESS
        a catch-up deal is applied automatically.
      parameters:
      - description: game UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: player
          schema:
            additionalProperties: true
            type: object
        "400":
          description: already in a game, game full, etc.
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: game not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Join a game
      tags:
      - players
...
  /games/{id}/players/{pid}/deal:
    post:
      consumes:
      - application/json
      description: Deals N cards from the shoe to a specific player. If the shoe has
        fewer than N cards remaining, all available cards are dealt (no error). Auto-ends
        the game if the shoe can no longer serve a full round.
      parameters:
      - description: game UUID
        in: path
        name: id
        required: true
        type: string
      - description: player UUID
        in: path
        name: pid
        required: true
        type: string
      - description: number of cards
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.dealRequest'
      produces:
      - application/json
      responses:
        "200":
          description: dealtCount + gameEnded
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deal cards to one player
      tags:
      - dealing
  /games/{id}/players/{pid}/hand:
    get:
      description: Returns the cards held by a player. Only the player themselves
        or an admin may view a hand.
      parameters:
      - description: game UUID
        in: path
        name: id
        required: true
        type: string
      - description: player UUID
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: cards array
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a player's hand
      tags:
      - dealing
...
securityDefinitions:
  BearerAuth:
    description: JWT obtained from POST /auth/register or POST /auth/login.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"